fix: close pipeline latency blind spots and drop-induced display freeze
Instrumentation and correctness work across the CAT-to-display path.

update_frequency_displays() committed its deduplication state before
queueing the subject update and ignored the return value. A dropped
enqueue therefore latched the value as already sent, freezing the
displayed frequency until the radio happened to retune. Commit the
dedup state only after a successful enqueue, and log the drop.

Subject queue ages stopped at dequeue, so everything from subject apply
through render and flush was unmeasured. Sample apply duration, render
duration, and oldest-enqueue to refresh-complete via LV_EVENT_RENDER_START,
LV_EVENT_RENDER_READY and LV_EVENT_REFR_READY, and report them in the
PIPELINE log. On DSI the flush DMA can still be in flight at REFR_READY,
so the end-to-end figure is a lower bound on panel-visible latency.

kenwood_if_data_t is 56 bytes and exceeded the 48-byte inline payload
limit, so every IF frame did a PSRAM malloc and free at the 6.7 Hz poll
rate. Raise the inline limit to 64 bytes, which costs 1 KB across the
64-slot queue, and static_assert that the struct stays inside it.

cat_polling_state_t::last_cat_activity held esp_timer_get_time() / 1000
in a uint32_t. Past 49.7 days of uptime the truncation combined with the
uint64 subtraction in cat_polling_is_cat_active() to report a permanent
timeout rather than a one-shot glitch. Widen the field and its accessor.

Also exposes UART RX ring occupancy, which is invisible to the queue age
stats because timestamps are only applied after bytes leave the ring, and
moves read_uart and cat_parser to core 1 above the LVGL task so core-0
network load cannot starve CAT ingest.

Not addressed: cat_polling_is_cat_active() still has no callers, so the
polling backoff it implements never runs. Wiring it up as-is would
deadlock, since the poll timers gate on the flag it clears and nothing
would then arrive to set it again.

+++ main/cat_polling.cpp
@@ -411,8 +411,8 @@ bool cat_polling_is_cat_active(void) {
     return g_polling_state.cat_connection_active;
 }
 
-uint32_t cat_polling_get_last_activity_time(void) {
-    return (uint32_t)g_polling_state.last_cat_activity;
+uint64_t cat_polling_get_last_activity_time(void) {
+    return g_polling_state.last_cat_activity;
 }
 
 const cat_polling_state_t* cat_polling_get_state(void) {
